refactor: move client to bundled directory
